style: reconstruct the code structure for npc
changed the code structure of npc according to NEMU
now support state, monitor, debug marco

+++ npc/include/memory.h
@@ -1,10 +1,11 @@
 #ifndef __NPC_MEMORY_H__
 #define __NPC_MEMORY_H__
 
+#include<common.h>
+
 #define MEM_MAX 0x8000000 
 #define MEM_BASE 0x80000000 
 
-#include <cstdint>
 static bool in_mem(uint32_t addr) { 
 	return addr - MEM_BASE < MEM_MAX; 
 }
@@ -13,6 +14,6 @@ extern "C" int pmem_read(int raddr);
 
 extern "C" void pmem_write(int waddr, int wdata, char wmask);
 
-int load_mem(const char* img);
+long load_mem(const char* img);
 
 #endif
